+++ src/main/java/co/com/system/invoice/persistence/sellingpoint/SellingPointRepository.java
@@ -2,5 +2,5 @@ package co.com.system.invoice.persistence.sellingpoint;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SellingPointRepository extends JpaRepository<SellingPointEntity, Long> {
+public interface SellingPointRepository extends JpaRepository<SellingPointEntity, Integer> {
 }
